+++ src/Unit.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Bakame\Tokei;
 
 use function intdiv;
-use function is_float;
 use function round;
 
 enum Unit
@@ -25,7 +24,7 @@ enum Unit
     private const int MICRO_PER_DAY = 24 * self::MICRO_PER_HOUR;
     private const int MICRO_PER_WEEK = 7 * self::MICRO_PER_DAY;
 
-    public function microseconds(): int
+    public function inMicroseconds(): int
     {
         return match ($this) {
             Unit::Week => self::MICRO_PER_WEEK,
@@ -40,7 +39,7 @@ enum Unit
 
     public function round(int $valueInMicro, RoundingMode $mode = RoundingMode::Nearest): int
     {
-        $unit = $this->microseconds();
+        $unit = $this->inMicroseconds();
 
         $precision = match ($mode) {
             RoundingMode::Floor => 0,
@@ -53,30 +52,28 @@ enum Unit
 
     public function toMicroseconds(int|float $value): int
     {
-        $res = $this->microseconds() * $value;
-
-        return is_float($res) ? (int) round($res) : $res;
+        return (int) round($this->inMicroseconds() * $value);
     }
 
     public function whole(int $value): int
     {
-        return intdiv($value, $this->microseconds());
+        return intdiv($value, $this->inMicroseconds());
     }
 
     public function remainder(int $value): int
     {
-        return $value % $this->microseconds();
+        return $value % $this->inMicroseconds();
     }
 
     public function divide(int $value): int|float
     {
-        return $value / $this->microseconds();
+        return $value / $this->inMicroseconds();
     }
 
     public function wrap(int $valueInMicro): int
     {
-        $unit = $this->microseconds();
+        $unit = $this->inMicroseconds();
 
-        return (($valueInMicro % $unit) + $unit) % $unit;
+        return ($valueInMicro % $unit + $unit) % $unit;
     }
 }